chore: refactored docs/
- Updated all README.adoc to README.mdx
- Moved `templates-md` and the contents to just `templates`
- Updated `hardhat.config.js` to use `docs/config`
- Updated `helpers.js` to account for adoc change
- Updated `docs/config.js` to use mdx
