--- contracts/access/README.adoc
@@ -1,12 +1,0 @@
-= Access
-
-[.readme-notice]
-NOTE: This document is better viewed at https://docs.openzeppelin.com/community-contracts/api/access
-
-This directory contains utility contracts to restrict access control in smart contracts. These include:
-
- * {AccessManagerLight}: A simpler version of an AccessManager that uses `bytes8` roles to allow function calls identified by their 4-bytes selector.
-
-== AccessManager
-
-{{AccessManagerLight}}
