+++ contracts/access/README.mdx
@@ -1,0 +1,15 @@
+---
+title: Access
+---
+
+<Callout>
+This document is better viewed at https://docs.openzeppelin.com/community-contracts/api/access
+</Callout>
+
+This directory contains utility contracts to restrict access control in smart contracts. These include:
+
+* AccessManagerLight: A simpler version of an AccessManager that uses `bytes8` roles to allow function calls identified by their 4-bytes selector.
+
+## AccessManager
+
+AccessManagerLight
