--- contracts/account/README.adoc
@@ -1,56 +1,0 @@
-= Account
-[.readme-notice]
-NOTE: This document is better viewed at https://docs.openzeppelin.com/community-contracts/api/account
-
-This directory includes contracts to build accounts for ERC-4337. These include:
-
- * {ERC7579Executor}: An executor module that enables executing calls from accounts where the it's installed.
- * {ERC7579DelayedExecutor}: An executor module that adds a delay before executing an account operation.
- * {ERC7579SelectorExecutor}: An executor module that restricts execution to specific function selectors.
- * {ERC7579Validator}: Abstract validator module for ERC-7579 accounts that provides base implementation for signature validation.
- * {ERC7579Signature}: Implementation of {ERC7579Validator} using ERC-7913 signature verification for address-less cryptographic keys and account signatures.
- * {ERC7579Multisig}: An extension of {ERC7579Validator} that enables validation using ERC-7913 signer keys.
- * {ERC7579MultisigWeighted}: An extension of {ERC7579Multisig} that allows different weights to be assigned to signers.
- * {ERC7579MultisigConfirmation}: An extension of {ERC7579Multisig} that requires each signer to provide a confirmation signature.
- * {ERC7579MultisigStorage}: An extension of {ERC7579Multisig} that allows storing presigned approvals in storage.
- * {PaymasterCore}: An ERC-4337 paymaster implementation that includes the core logic to validate and pay for user operations.
- * {PaymasterERC20}: A paymaster that allows users to pay for user operations using ERC-20 tokens.
- * {PaymasterERC20Guarantor}: A paymaster that enables third parties to guarantee user operations by pre-funding gas costs, with the option for users to repay or for guarantors to absorb the cost.
- * {PaymasterERC721Owner}: A paymaster that allows users to pay for user operations based on ERC-721 ownership.
- * {PaymasterSigner}: A paymaster that allows users to pay for user operations using an authorized signature.
-
-== Modules
-
-=== Executors
-
-{{ERC7579Executor}}
-
-{{ERC7579DelayedExecutor}}
-
-{{ERC7579SelectorExecutor}}
-
-=== Validators
-
-{{ERC7579Validator}}
-
-{{ERC7579Signature}}
-
-{{ERC7579Multisig}}
-
-{{ERC7579MultisigWeighted}}
-
-{{ERC7579MultisigConfirmation}}
-
-{{ERC7579MultisigStorage}}
-
-== Paymaster
-
-{{PaymasterCore}}
-
-{{PaymasterERC20}}
-
-{{PaymasterERC20Guarantor}}
-
-{{PaymasterERC721Owner}}
-
-{{PaymasterSigner}}
