--- contracts/crosschain/README.adoc
@@ -1,27 +1,0 @@
-= Crosschain
-
-[.readme-notice]
-NOTE: This document is better viewed at https://docs.openzeppelin.com/community-contracts/api/crosschain
-
-Gateways are contracts that enable cross-chain communication. These can either be a message source or a destination according to ERC-7786.
-
- * {ERC7786Receiver}: ERC-7786 cross-chain message receiver.
- * {ERC7786OpenBridge}: ERC-7786 "N out of M" gateway. Sends a message through M gateways and executes on the destination if N received it.
-
-Developers can access interoperability protocols through gateway adapters. The library includes the following gateway adapters:
-
- * {AxelarGatewayAdapter}: ERC-7786 gateway adapter for Axelar.
-
-== Gateways
-
-{{ERC7786OpenBridge}}
-
-== Clients
-
-{{ERC7786Receiver}}
-
-== Adapters
-
-=== Axelar
-
-{{AxelarGatewayAdapter}}
