--- contracts/governance/README.adoc
@@ -1,12 +1,0 @@
-= Governance
-
-[.readme-notice]
-NOTE: This document is better viewed at https://docs.openzeppelin.com/contracts/api/governance
-
-This directory includes extensions and utilities for on-chain governance.
-
-* {TimelockControllerEnumerable}: Extension of OpenZeppelin's TimelockController with enumerable operations support.
-
-== Timelock
-
-{{TimelockControllerEnumerable}}
