+++ contracts/interfaces/README.mdx
@@ -1,0 +1,31 @@
+---
+title: Interfaces
+---
+
+<Callout>
+This document is better viewed at https://docs.openzeppelin.com/contracts/api/interfaces
+</Callout>
+
+## List of standardized interfaces
+
+These interfaces are available as `.sol` files. These are useful to interact with third party contracts that implement them.
+
+* IERC7786GatewaySource, IERC7786Receiver
+* IERC7802
+* IERC7821
+* IERC7913SignatureVerifier
+* IERC7943
+
+## Detailed ABI
+
+IERC7786GatewaySource
+
+IERC7786Receiver
+
+IERC7802
+
+IERC7821
+
+IERC7913SignatureVerifier
+
+IERC7943
