--- contracts/proxy/README.adoc
@@ -1,12 +1,0 @@
-= Proxy
-
-[.readme-notice]
-NOTE: This document is better viewed at https://docs.openzeppelin.com/community-contracts/proxy
-
-Variants of proxy patterns, which are contracts that allow to forward a call to an implementation contract by using `delegatecall`. This contracts include:
-
- * {HybridProxy}: An ERC-1967 proxy that uses the implementation slot as a beacon in a way that a user can upgrade to an implementation of their choice.
-
-== General
-
-{{HybridProxy}}
