+++ contracts/proxy/README.mdx
@@ -1,0 +1,15 @@
+---
+title: Proxy
+---
+
+<Callout>
+This document is better viewed at https://docs.openzeppelin.com/community-contracts/proxy
+</Callout>
+
+Variants of proxy patterns, which are contracts that allow to forward a call to an implementation contract by using `delegatecall`. This contracts include:
+
+* HybridProxy: An ERC-1967 proxy that uses the implementation slot as a beacon in a way that a user can upgrade to an implementation of their choice.
+
+## General
+
+HybridProxy
