+++ contracts/token/README.mdx
@@ -1,0 +1,32 @@
+---
+title: Tokens
+---
+
+<Callout>
+This document is better viewed at https://docs.openzeppelin.com/community-contracts/token
+</Callout>
+
+Set of extensions and utilities for tokens (e.g ERC-20, ERC-721, ERC-1155) and derivated ERCs (e.g. ERC-4626, ERC-1363).
+
+* OnTokenTransferAdapter: Adapter of the ERC-1363 receiver interface to comply with Chainlink’s 667 interface.
+* ERC20Allowlist: Extension of ERC20 with transfers and approvals that require users to be registered into an allowlist.
+* ERC20Blocklist: Extension of ERC20 with transfers and approvals that can be disabled by adding users into a blocklist.
+* ERC20Collateral: Oracle-agnostic extension of ERC20 that limits the total supply based on a collateral amount.
+* ERC20Custodian: Extension of ERC20 that implements an access-control agnostic approach to define a custodian that can freeze user’s transfers and approvals.
+* ERC4626Fees: ERC4626 vault with fees on entry (deposit/mint) or exit (withdraw/redeem).
+
+## General
+
+OnTokenTransferAdapter
+
+## ERC20
+
+ERC20Allowlist
+
+ERC20Blocklist
+
+ERC20Collateral
+
+ERC20Custodian
+
+ERC4626Fees
