--- contracts/utils/README.adoc
@@ -1,19 +1,0 @@
-= Utilities
-
-[.readme-notice]
-NOTE: This document is better viewed at https://docs.openzeppelin.com/community-contracts/utils
-
-Miscellaneous contracts and libraries containing utility functions you can use to improve security, work with new data types, or safely use low-level primitives.
-
- * {EnumerableSetExtended} and {EnumerableMapExtended}: Extensions of the `EnumerableSet` and `EnumerableMap` libraries with more types, including non-value types.
- * {Masks}: Library to handle `bytes32` masks.
-
-== Structs
-
-{{EnumerableSetExtended}}
-
-{{EnumerableMapExtended}}
-
-== Libraries
-
-{{Masks}}
