+++ contracts/utils/README.mdx
@@ -1,0 +1,22 @@
+---
+title: Utilities
+---
+
+<Callout>
+This document is better viewed at https://docs.openzeppelin.com/community-contracts/utils
+</Callout>
+
+Miscellaneous contracts and libraries containing utility functions you can use to improve security, work with new data types, or safely use low-level primitives.
+
+* EnumerableSetExtended and EnumerableMapExtended: Extensions of the `EnumerableSet` and `EnumerableMap` libraries with more types, including non-value types.
+* Masks: Library to handle `bytes32` masks.
+
+## Structs
+
+EnumerableSetExtended
+
+EnumerableMapExtended
+
+## Libraries
+
+Masks
