--- contracts/utils/cryptography/README.adoc
@@ -1,33 +1,0 @@
-= Cryptography
-
-[.readme-notice]
-NOTE: This document is better viewed at https://docs.openzeppelin.com/community-contracts/utils/cryptography
-
-A collection of contracts and libraries that implement various signature validation schemes and cryptographic primitives. These utilities enable secure authentication, multisignature operations, and advanced cryptographic operations in smart contracts.
-
- * {ZKEmailUtils}: Library for ZKEmail signature validation utilities, enabling email-based authentication through zero-knowledge proofs.
- * {WebAuthn}: Library for verifying WebAuthn Authentication Assertions.
- * {DKIMRegistry}: Implementation of https://eips.ethereum.org/EIPS/eip-7969[ERC-7969] to enable onchain verification of DomainKeys Identified Mail (DKIM) signatures.
- * {SignerZKEmail}: Implementation of an https://docs.openzeppelin.com/contracts/5.x/api/utils/cryptography#AbstractSigner[AbstractSigner] that enables email-based authentication through zero-knowledge proofs.
- * {SignerWebAuthn}: Implementation of https://docs.openzeppelin.com/contracts/5.x/api/utils/cryptography#SignerP256[SignerP256] that supports WebAuthn authentication assertions.
- * {ERC7913ZKEmailVerifier}, {ERC7913WebAuthnVerifier}: Ready to use ERC-7913 signature verifiers for ZKEmail and WebAuthn.
-
-== Utils
-
-{{ZKEmailUtils}}
-
-{{WebAuthn}}
-
-{{DKIMRegistry}}
-
-== Abstract Signers
-
-{{SignerZKEmail}}
-
-{{SignerWebAuthn}}
-
-== Verifiers
-
-{{ERC7913ZKEmailVerifier}}
-
-{{ERC7913WebAuthnVerifier}}
