+++ docs/config.js
@@ -6,14 +6,13 @@ module.exports = {
   outputDir: 'docs/modules/api/pages',
   templates: 'docs/templates',
   exclude: ['mocks'],
-  pageExtension: '.adoc',
+  pageExtension: '.mdx',
   pages: (_, file, config) => {
-    // For each contract file, find the closest README.adoc and return its location as the output page path.
     const sourcesDir = path.resolve(config.root, config.sourcesDir);
     let dir = path.resolve(config.root, file.absolutePath);
     while (dir.startsWith(sourcesDir)) {
       dir = path.dirname(dir);
-      if (fs.existsSync(path.join(dir, 'README.adoc'))) {
+      if (fs.existsSync(path.join(dir, 'README.mdx'))) {
         return path.relative(sourcesDir, dir) + config.pageExtension;
       }
     }
