--- docs/templates-md/page.hbs
@@ -1,13 +1,0 @@
----
-title: "{{title}}"
-description: "{{description}}"
----
-
-{{#with-prelude}}
-{{readme (readme-path)}}
-{{/with-prelude}}
-
-{{#each items}}
-{{>contract}}
-
-{{/each}}
