+++ docs/templates/page.hbs
@@ -1,4 +1,13 @@
-:github-icon: pass:[<svg class="icon"><use href="#github-icon"/></svg>]
+---
+title: "{{title}}"
+description: "{{description}}"
+---
+
 {{#with-prelude}}
 {{readme (readme-path)}}
 {{/with-prelude}}
+
+{{#each items}}
+{{>contract}}
+
+{{/each}}
